Remove a line that was causing edoc problems.
This commit drops a comment that was causing edoc to fail to
build the documentation, with the following error:

./src/p1_options.erl, in module header: at line 3:
multiple '<...>' sections.

The line had duplicate information since the creation date is in
git history and the author is already noted in the line above.
Thus, it was simplest to remove the line.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>